update CosaUtilGetIfAddr() result data type ( ULONG -> uint32_t )
An IPv4 address is a 32bit value.

+++ source/TR-181/middle_layer_src/cosa_apis_util.c
@@ -131,7 +131,7 @@ CosaUtilStringToHex
     return ANSC_STATUS_SUCCESS;
 }
 
-ULONG CosaUtilGetIfAddr (char *netdev)
+uint32_t CosaUtilGetIfAddr (char *netdev)
 {
     struct ifreq ifr;
     struct sockaddr_in *ipaddr;
